Created the rest of the delegates

+++ CIS560-Project/CIS560-Project/Controllers/ITrainingRunRepository.cs
@@ -13,6 +13,6 @@ namespace CIS560_Project.Controllers
 
         IReadOnlyList<TrainingRun> FetchTrainingRuns(int runnerId);
 
-        TrainingRun CreateTrainingRun(int runnerId, DateTime date, int distance, double time, double averageHeartRate = 0.0, bool isArchived = false);
+        TrainingRun CreateTrainingRun(int runnerId, DateTime date, int distance, int time, double averageHeartRate = 0.0, bool isArchived = false);
     }
 }
